- updated fetch_survey and response_parsers
- Updated the code to include the the 'matrix of drop-down menus' answers. In order to make it work, two additional changes were made:
(i) include a nested loop in fetch_survey.R to capture by each column the choices (capture column name in answers and drop-down choices in choices).
(ii) changed process_matrix in response_parsers.R to include the 'menu' subtype. In this type, the name of the column is a combination of three variables: question_id, row_id and col_id.